feat: propagate action vernacular
